% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlzMWRdate.R
\name{anlzMWRdate}
\alias{anlzMWRdate}
\title{Analyze trends by date in results file}
\usage{
anlzMWRdate(
  res,
  param,
  acc,
  thresh,
  group = c("site", "all"),
  threshcol = "tan",
  site = NULL,
  resultatt = NULL,
  dtrng = NULL,
  ptsize = 2,
  repel = TRUE,
  labsize = 3,
  yscl = c("auto", "log", "linear"),
  runchk = TRUE,
  warn = TRUE
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{param}{character string of the parameter to plot, must conform to entries in the \code{"Simple Parameter"} column of \code{\link{paramsMWR}}}

\item{acc}{character string of path to the data quality objectives file for accuracy or \code{data.frame} returned by \code{\link{readMWRacc}}}

\item{thresh}{character indicating if relevant freshwater or marine threshold lines are included, one of \code{"fresh"}, \code{"marine"}, or \code{"none"}}

\item{group}{character indicating whether the results are grouped by site (default) or combined across all sites}

\item{threshcol}{character indicating color of threshold lines if available}

\item{site}{character string of sites to include, default all}

\item{resultatt}{character string of result attributes to plot, default all}

\item{dtrng}{character string of length two for the date ranges as YYYY-MM-DD, optional}

\item{ptsize}{numeric indicating size of the points}

\item{repel}{logical indicating if overlapping site labels are offset}

\item{labsize}{numeric indicating font size for the site labels, only if \code{group = "site"}}

\item{yscl}{character indicating one of \code{"auto"} (default), \code{"log"}, or \code{"linear"}, see details}

\item{runchk}{logical to run data checks with \code{\link{checkMWRresults}} or \code{\link{checkMWRacc}}, applies only if \code{res} or \code{acc} are file paths}

\item{warn}{logical to return warnings to the console (default)}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified.
}
\description{
Analyze trends by date in results file
}
\details{
Results are shown for the selected parameter as continuous line plots over time. Specifying \code{group = "site"} plot a separate line for each site.  Specifying \code{group = "all"} will average results across sites for each date.

Threshold lines applicable to marine or freshwater environments can be included in the plot by using the \code{thresh} argument.  These thresholds are specific to each parameter and can be found in the \code{\link{thresholdMWR}} file.  Threshold lines are plotted only for those parameters with entries in \code{\link{thresholdMWR}} and only if the value in \code{`Result Unit`} matches those in \code{\link{thresholdMWR}}. The threshold lines can be suppressed by setting \code{thresh = 'none'}.

The y-axis scaling as arithmetic (linear) or logarithmic can be set with the \code{yscl} argument.  If \code{yscl = "auto"} (default), the scaling is  determined automatically from the data quality objective file for accuracy, i.e., parameters with "log" in any of the columns are plotted on log10-scale, otherwise arithmetic. Setting \code{yscl = "linear"} or \code{yscl = "log"} will set the axis as linear or log10-scale, respectively, regardless of the information in the data quality objective file for accuracy.

Any entries in \code{resdat} in the \code{"Result Value"} column as \code{"BDL"} or \code{"AQL"} are replaced with appropriate values in the \code{"Quantitation Limit"} column, if present, otherwise the \code{"MDL"} or \code{"UQL"} columns from the data quality objectives file for accuracy are used.  Values as \code{"BDL"} use one half of the appropriate limit.
}
\examples{
# results data path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# results data
resdat <- readMWRresults(respth)

# accuracy path
accpth <- system.file('extdata/ExampleDQOAccuracy.xlsx', 
     package = 'MassWateR')

# accuracy data
accdat <- readMWRacc(accpth)

# select sites
anlzMWRdate(res = resdat, param = 'DO', acc = accdat, group = 'site', thresh = 'fresh',
     site = c("ABT-026", "ABT-077"))

# combined sites
anlzMWRdate(res = resdat, param = 'DO', acc = accdat, group = 'all', thresh = 'fresh',
     site = c("ABT-026", "ABT-077"))

}
