% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlzMWRoutlier.R
\name{anlzMWRoutlier}
\alias{anlzMWRoutlier}
\title{Analyze outliers in results file}
\usage{
anlzMWRoutlier(
  res,
  param,
  type = c("month", "site"),
  threshold = FALSE,
  outliertxt = FALSE
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{param}{character string of the parameter to plot, must conform entries in the \code{"Simple Parameter"} of \code{\link{paramsMWR}}}

\item{type}{character indicating whether the summaries are grouped by month (default) or site}

\item{threshold}{logical indicating if threshold lines are applied to the plot from \code{\link{thresholdMWR}}, if available}

\item{outliertxt}{logical indicating if outliers are returned to the console after plotting}
}
\value{

}
\description{
Analyze outliers in results file
}
\examples{
NULL
}
