% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_acc.R
\encoding{UTF-8}
\name{check_acc}
\alias{check_acc}
\title{Check data quality objective accuracy data}
\usage{
check_acc(accdat)
}
\arguments{
\item{accdat}{input data frame}
}
\value{
\code{accdat} is returned as is if no errors are found, otherwise an informative error message is returned prompting the user to make the required correction to the raw data before proceeding.
}
\description{
Check data quality objective accuracy data
}
\details{
This function is used internally within \code{\link{read_acc}} to run several checks on the input data for completeness and conformance to WQX requirements

The following checks are made:
\itemize{
\item{Column name spelling: }{Should be the following: Parameter, uom, MDL, UQL, Value Range, Field Duplicate, Lab Duplicate, Field Blank, Lab Blank, Spike/Check Accuracy}
\item{Columns present: }{All columns from the previous check should be present}
\item{Non-numeric values in MDL, UQL: }{Values entered in columns MDL and UQL should be numeric}
\item{Unrecognized characters: }{Fields describing accuracy checks should not include symbols or text other than \eqn{<=}, \eqn{\leq}, \eqn{<}, \eqn{>=}, \eqn{\geq}, \eqn{>}, \eqn{\pm}, \%, BDL, AQL, log, or all}
\item{Parameter: }{Should match parameter names in the \code{Simple Parameter} or \code{WQX Parameter} columns of the \code{\link{params}} data}
\item{Units: }{No missing entries in units (\code{uom}), except pH which can be blank}
\item{Single unit: }{Each unique \code{Parameter} should have only one type for the units (\code{uom})}
\item{Correct units: }{Each unique \code{Parameter} should have an entry in the units (\code{uom}) that matches one of the acceptable values in the \code{Units of measure} column of the \code{\link{params}} data}
}
}
\examples{
accpth <- system.file('extdata/ExampleDQOAccuracy_final.xlsx', package = 'MassWateR')

accdat <- readxl::read_excel(accpth, na = c('NA', 'na', '')) 
      
check_acc(accdat)
}
