% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dqoaccuracy.R
\encoding{UTF-8}
\name{check_dqoaccuracy}
\alias{check_dqoaccuracy}
\title{Check data quality objective accuracy data}
\usage{
check_dqoaccuracy(dqoaccdat)
}
\arguments{
\item{dqoaccdat}{input data frame}
}
\value{
\code{dqoaccdat} is returned as is if no errors are found, otherwise an informative error message is returned prompting the user to make the required correction to the raw data before proceeding.
}
\description{
Check data quality objective accuracy data
}
\details{
This function is used internally within \code{\link{read_dqoaccuracy}} to run several checks on the input data for completeness and conformance to WQX requirements

The following checks are made:
\itemize{
\item{Column name spelling: }{Should be the following: Parameter, uom, MDL, UQL, Value Range, Field Duplicate, Lab Duplicate, Field Blank, Lab Blank, Spike/Check Accuracy}
\item{Columns present: }{All columns from the previous check should be present}
\item{Non-numeric values in MDL, UQL: }{Values entered in columns MDL and UQL should be numeric}
\item{Unrecognized characters: }{Fields describing accuracy checks should not include symbols or text other than <=, ≤, <, >=, ≥, >, ±, \%, BDL, AQL, log, or all}
}
}
\examples{
dqoaccpth <- system.file('extdata/ExampleDQOAccuracy_final.xlsx', package = 'MassWateR')

dqoaccdat <- readxl::read_excel(dqoaccpth, na = c('NA', 'na', '')) 
      
check_dqoaccuracy(dqoaccdat)
}
