% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dqocompleteness.R
\name{check_dqocompleteness}
\alias{check_dqocompleteness}
\title{Check data quality objective completeness data}
\usage{
check_dqocompleteness(dat)
}
\arguments{
\item{dat}{input data frame}
}
\value{
\code{dat} is returned as is if no errors are found, otherwise an informative error message is returned prompting the user to make the required correction to the raw data before proceeding.
}
\description{
Check data quality objective completeness data
}
\details{
This function is used internally within \code{\link{read_dqocompleteness}} to run several checks on the input data for completeness and conformance to WQX requirements

The following checks are made:
\itemize{
\item{Column name spelling: }{Should be the following: Parameter, Field Duplicate, Lab Duplicate, Field Blank, Lab Blank, Spike/Check Accuracy, \% Completeness}
\item{Columns present: }{All columns from the previous check should be present}
\item{Non-numeric values: }{Values entered in columns other than the first should be numeric}
\item{Values outside of 0 - 100: }{Values entered in columns other than the first should not be outside of 0 and 100}
}
}
\examples{
library(dplyr)

pth <- system.file('extdata/ExampleDQOCompleteness_final.xlsx', package = 'MassWateR')

dat <- suppressMessages(readxl::read_excel(pth, 
      skip = 1, na = c('na', ''), 
      col_types = c('text', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric')
    )) \%>\% 
    rename(`\% Completeness` = `...7`)
    
check_dqocompleteness(dat)
}
