% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_results}
\alias{check_results}
\title{Check water quality monitoring results}
\usage{
check_results(resdat)
}
\arguments{
\item{resdat}{input data frame for results}
}
\value{
\code{resdat} is returned as is if no errors are found, otherwise an informative error message is returned prompting the user to make the required correction to the raw data before proceeding.
}
\description{
Check water quality monitoring results
}
\details{
This function is used internally within \code{\link{read_results}} to run several checks on the input data for completeness and conformance to WQX requirements

The following checks are made:
\itemize{
\item{Column name spelling: }{Should be the following: Monitoring Location ID, Activity Type, Activity Start Date, Activity Start Time, Activity Depth/Height Measure, Activity Depth/Height Unit, Activity Relative Depth Name, Characteristic Name, Result Value, Result Unit, Quantitation Limit, QC Reference Value, Result Measure Qualifier, Result Attribute.}
\item{Columns present: }{All columns from the previous check should be present, Result Attribute is optional}
\item{Activity Type: }{Should be one of Field Msr/Obs, Sample-Routine, Quality Control Sample-Field Blank, Quality Control Sample-Lab Blank, Quality Control Sample-Lab Duplicate, Quality Control Sample-Lab Spike}
\item{Date formats: }{Should be mm/dd/yyyy and parsed correctly on import}
\item{Time formats: }{Should be HH:MM and parsed correctly on import}
\item{Activity Relative Depth Name: }{Should be either Surface, Bottom, < 1m / 3.3ft or blank}
\item{Characteristic Name: }{Should match parameter names in the \code{Simple Parameter} or \code{WQX Parameter} columns of the \code{\link{params}} data}
\item{Result Value: }{Should be a numeric value or a text value as AQL or BDL}
\item{QC Reference Value: }{Should be a numeric value or a text value as AQL or BDL}
\item{Result Unit: }{No missing entries in \code{Result Unit}, except pH which can be blank}
\item{Single Result Unit: }{Each unique parameter in \code{Characteristic Name} should have only one entry in \code{Result Unit}}
\item{Correct Result Unit: }{Each unique parameter in \code{Characteristic Name} should have an entry in \code{Result Unit} that matches one of the acceptable values in the \code{Units of measure} column of the \code{\link{params}} data}
}
}
\examples{
respth <- system.file('extdata/ExampleResults_final.xlsx', package = 'MassWateR')

resdat <- readxl::read_excel(respth, 
  col_types = c('text', 'text', 'date', 'date', 'text', 'text', 'text', 'text', 'text', 'text', 
             'text', 'text', 'text', 'text'))
             
check_results(resdat)
}
