% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqo.R
\docType{data}
\name{exdqo}
\alias{exdqo}
\title{Example file of data quality objectives}
\format{
A \code{data.frame} of 21 rows and 10 columns
}
\usage{
exdqo
}
\description{
Example file of data quality objectives
}
\details{
An example of data quality objectives formatted for comparison with data in \code{\link{exresults}} in the QAQC or exploratory analysis workflow
}
\examples{
\dontrun{
library(readxl)

exdqo <- read_excel('inst/extdata/ExampleDQO_final.xlsx')

save(exdqo, file = 'data/exdqo.RData', compress = 'xz')
}
}
\keyword{datasets}
