% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqoacc.R
\docType{data}
\name{exdqoacc}
\alias{exdqoacc}
\title{Example file of data quality objectives for accuracy}
\format{
A \code{data.frame} of 21 rows and 10 columns
}
\usage{
exdqoacc
}
\description{
Example file of data quality objectives for accuracy
}
\details{
An example of data quality objectives for accuracy formatted for comparison with data in \code{\link{exresults}} in the QAQC or exploratory analysis workflow
}
\examples{
\dontrun{
library(readxl)

exdqoacc <- read_excel('inst/extdata/ExampleDQOAccuracy_final.xlsx', sheet = 'Accuracy', na = 'na')

save(exdqoacc, file = 'data/exdqoacc.RData', compress = 'xz')
}
}
\keyword{datasets}
