% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdqocom.R
\docType{data}
\name{exdqocom}
\alias{exdqocom}
\title{Example file of data quality objectives for completeness}
\format{
A \code{data.frame} of 21 rows and 10 columns
}
\usage{
exdqocom
}
\description{
Example file of data quality objectives for completeness
}
\details{
An example of data quality objectives for completeness formatted for comparison with data in \code{\link{exresults}} in the QAQC or exploratory analysis workflow
}
\examples{
\dontrun{
library(readxl)
library(dplyr)

exdqocom <- read_excel('inst/extdata/ExampleDQOCompleteness_final.xlsx', sheet = 'Completeness', skip = 1, na = 'na') \%>\% 
  rename(`\% Completeness` = `...7`)

save(exdqocom, file = 'data/exdqocom.RData', compress = 'xz')
}
}
\keyword{datasets}
