% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exresults.R
\docType{data}
\name{exresults}
\alias{exresults}
\title{Example file of water quality results}
\format{
A \code{data.frame} of 2584 rows and 14 columns
}
\usage{
exresults
}
\description{
Example file of water quality results
}
\details{
An example of water quality data formatted for input in the QAQC or exploratory analysis workflow
}
\examples{
\dontrun{
library(readxl)

exresults <- read_excel('inst/extdata/ExampleResults_final.xlsx')

save(exresults, file = 'data/exresults.RData', compress = 'xz')
}
}
\keyword{datasets}
