% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_acc.R
\name{form_acc}
\alias{form_acc}
\title{Format data quality objective accuracy data}
\usage{
form_acc(accdat)
}
\arguments{
\item{accdat}{input data fram}
}
\description{
Format data quality objective accuracy data
}
\details{
This function is used internally within \code{\link{read_acc}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item{Convert Parameter: }{All parameters are converted to \code{Simple Parameter} in \code{\link{params}} as needed}
}
}
\examples{
accpth <- system.file('extdata/ExampleDQOAccuracy_final.xlsx', package = 'MassWateR')

accdat <- readxl::read_excel(accpth, na = c('NA', 'na', '')) 

form_acc(accdat)
}
