% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_dqoaccuracy.R
\name{form_dqoaccuracy}
\alias{form_dqoaccuracy}
\title{Format data quality objective accuracy data}
\usage{
form_dqoaccuracy(dqoaccdat)
}
\arguments{
\item{dqoaccdat}{input data fram}
}
\description{
Format data quality objective accuracy data
}
\details{
This function is used internally within \code{\link{read_dqoaccuracy}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item{Convert Parameter: }{All parameters are converted to \code{Simple Parameter} in \code{\link{params}} as needed}
}
}
\examples{
dqoaccpth <- system.file('extdata/ExampleDQOAccuracy_final.xlsx', package = 'MassWateR')

dqoaccdat <- readxl::read_excel(dqoaccpth, na = c('NA', 'na', '')) 

form_dqoaccuracy(dqoaccdat)
}
