% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_dqocompleteness.R
\name{form_dqocompleteness}
\alias{form_dqocompleteness}
\title{Format data quality objective completeness data}
\usage{
form_dqocompleteness(dqocomdat)
}
\arguments{
\item{dqocomdat}{input data fram}
}
\description{
Format data quality objective completeness data
}
\details{
This function is used internally within \code{\link{read_dqocompleteness}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item{Convert Parameter: }{All parameters are converted to \code{Simple Parameter} in \code{\link{params}} as needed}
}
}
\examples{
library(dplyr)

dqocompth <- system.file('extdata/ExampleDQOCompleteness_final.xlsx', package = 'MassWateR')

dqocomdat <- suppressMessages(readxl::read_excel(dqocompth, 
      skip = 1, na = c('NA', 'na', ''), 
      col_types = c('text', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric', 'numeric')
    )) \%>\% 
    rename(`\% Completeness` = `...7`)
    
form_dqocompleteness(dqocomdat)
}
