% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_results.R
\name{form_results}
\alias{form_results}
\title{Format water quality monitoring results}
\usage{
form_results(resdat, tzone = "America/Jamaica")
}
\arguments{
\item{resdat}{input data frame for results}

\item{tzone}{character string for time zone}
}
\description{
Format water quality monitoring results
}
\details{
This function is used internally within \code{\link{read_results}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item{Fix date and time inputs: }{Activty Start Date is converted to YYYY-MM-DD as a date object, Actvity Start Time is convered to HH:MM as a character to fix artifacts from Excel import},
\item{Create duplicate rows for entries in QC Reference Value: }{This is done if a value is found in the QC Reference Value column by creating a “new” row with the same location, date, time, etc as the original but with an appropriate quality control entry for the Activity Type}
\item{Remove non-text or math symbols from Result Unit: }{Values are stripped, e.g., the degree symbols for degrees C}
}
}
\examples{
respth <- system.file('extdata/ExampleResults_final.xlsx', package = 'MassWateR')

resdat <- readxl::read_excel(respth, 
  col_types = c('text', 'text', 'date', 'date', 'text', 'text', 'text', 'text', 'text', 'text', 
             'text', 'text', 'text', 'text'))
form_results(resdat)
}
