% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{params}
\alias{params}
\title{Master parameter list and units for Characteristic Name column in results data}
\format{
A \code{data.frame} of 39 rows and 5 columns
}
\usage{
params
}
\description{
Master parameter list and units for Characteristic Name column in results data
}
\details{
This information is used to verify the correct format of input results data in \code{\link{form_results}}.  A column showing the corresponding WQX names is also included.
}
\examples{
\dontrun{
library(readxl)
library(dplyr)

params <- read_excel('inst/extdata/Parameter.Mapping_4-29-22.xlsx') \%>\% 
  rename(Note = `...5`)

save(params, file = 'data/params.RData', compress = 'xz')
}
}
\keyword{datasets}
