% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acc.R
\name{read_acc}
\alias{read_acc}
\title{Read data quality objectives for accuracy from an external file}
\usage{
read_acc(accpth, runchk = TRUE)
}
\arguments{
\item{accpth}{character string of path to the data quality objectives file for accuracy}

\item{runchk}{logical to run data checks with \code{\link{check_acc}}}
}
\value{
A formatted data frame of data quality objectives for completeness that can be used for downstream analysis
}
\description{
Read data quality objectives for accuracy from an external file
}
\details{
Data are imported with \code{\link[readxl]{read_excel}} and checked with \code{\link{check_acc}}.
}
\examples{
accpth <- system.file('extdata/ExampleDQOAccuracy_final.xlsx', package = 'MassWateR')

accdat <- read_acc(accpth)
head(accdat)
}
