% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dqoaccuracy.R
\name{read_dqoaccuracy}
\alias{read_dqoaccuracy}
\title{Read data quality objectives for accuracy from an external file}
\usage{
read_dqoaccuracy(pth, runchk = TRUE)
}
\arguments{
\item{pth}{character string of path to the file}

\item{runchk}{logical to run data checks with \code{\link{check_dqoaccuracy}}}
}
\value{
A formatted data frame of data quality objectives for completeness that can be used for downstream analysis
}
\description{
Read data quality objectives for accuracy from an external file
}
\details{
Data are imported with \code{\link[readxl]{read_excel}} and checked with \code{\link{check_dqoaccuracy}}.
}
\examples{
pth <- system.file('extdata/ExampleDQOAccuracy_final.xlsx', package = 'MassWateR')

dqoaccuracy <- read_dqoaccuracy(pth)
head(dqoaccuracy)
}
