% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_frecom.R
\name{read_frecom}
\alias{read_frecom}
\title{Read data quality objectives for completeness from an external file}
\usage{
read_frecom(frecompth, runchk = TRUE)
}
\arguments{
\item{frecompth}{character string of path to the data quality objectives file for completeness}

\item{runchk}{logical to run data checks with \code{\link{check_frecom}}}
}
\value{
A formatted data frame of data quality objectives for completeness that can be used for downstream analysis
}
\description{
Read data quality objectives for completeness from an external file
}
\details{
Data are imported with \code{\link[readxl]{read_excel}} and checked with \code{\link{check_frecom}}.
}
\examples{
frecompth <- system.file('extdata/ExampleDQOFrequencyCompleteness_final.xlsx', 
     package = 'MassWateR')

frecomdat <- read_frecom(frecompth)
head(frecomdat)
}
