% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results}
\alias{read_results}
\title{Read water quality monitoring results from an external file}
\usage{
read_results(respth, runchk = TRUE, tzone = "America/Jamaica")
}
\arguments{
\item{respth}{character string of path to the results file}

\item{runchk}{logical to run data checks with \code{\link{check_results}}}

\item{tzone}{character string for time zone, passed to \code{\link{form_results}}}
}
\value{
A formatted water quality monitoring results data frame that can be used for downstream analysis
}
\description{
Read water quality monitoring results from an external file
}
\details{
Date are imported with \code{\link[readxl]{read_excel}}, checked with \code{\link{check_results}}, and formatted with \code{\link{form_results}}.
}
\examples{
respth <- system.file('extdata/ExampleResults_final.xlsx', package = 'MassWateR')

resdat <- read_results(respth)
head(resdat)
}
