% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMWRresultspivot.R
\name{readMWRresultspivot}
\alias{readMWRresultspivot}
\title{Create summary tables of the water quality monitoring results}
\usage{
readMWRresultspivot(respth, table)
}
\arguments{
\item{respth}{character string of path to the results file}

\item{table}{character string indicating which summary table to create, see details}
}
\value{
A summary table of counts of observations in the water quality monitoring result for the selected table
}
\description{
Create summary tables of content in the water quality monitoring results to troubleshoot checks with \code{\link{readMWRresults}}
}
\details{
Acceptable options for the \code{table} argument include \code{'parameter, unit'} or \code{'activity type, parameter'}.  Both options return summarized tables that show the unique counts of observations for the column combination provied by \code{table}.  For example, \code{table = 'parameter, unit'} will return a table where the rows are the unique parameters (\code{"Characteristic Name"}) and the columns are the unique units (\code{"Result Unit"}) found in the results file. The numbers in the table show the unique observations that correspond to each parameter and result unit combination. Similarly, \code{table = 'activity type, parameter'} will return a table where the rows are the unique activity types (\code{"Activity Type"}) and the columns for the unique parameters (\code{"Characteristic Name"}).

These tables can be useful to troubleshoot the checks when importing the water quality monitoring result file with \code{readMWRresults}, e.g., by seeing which units apply to a parameter and if they correspond to the acceptable values for the package (see \href{https://massbays-tech.github.io/MassWateR/articles/MassWateR.html#data-import-and-checks}{https://massbays-tech.github.io/MassWateR/articles/MassWateR.html#data-import-and-checks}).
}
\examples{
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# parameters and units
readMWRresultspivot(respth, table = c('parameter, unit'))

# activity types and parameters
readMWRresultspivot(respth, table = c('activity type, parameter'))
}
