% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMWRresultsview.R
\name{readMWRresultsview}
\alias{readMWRresultsview}
\title{Create summary views of the water quality monitoring results}
\usage{
readMWRresultsview(respth, columns = NULL)
}
\arguments{
\item{respth}{character string of path to the results file}

\item{columns}{character string indicating which columns to view, defaults to all}
}
\value{
Data viewers (from \code{\link[utils]{View}}) for the selected columns, one per column showing the unique values in the column.
}
\description{
Create summary views of unique values in columns in the water quality monitoring results to troubleshoot checks with \code{\link{readMWRresults}}
}
\details{
Acceptable options for the \code{columns} argument include any of the column names in the results file. The default setting (\code{NULL}) will open data viewers for every column in the results file (fourteen windows total).

The output of this function can be useful to troubleshoot the checks when importing the water quality monitoring result file with \code{readMWRresults} (see \href{https://massbays-tech.github.io/MassWateR/articles/MassWateR.html#data-import-and-checks}{https://massbays-tech.github.io/MassWateR/articles/MassWateR.html#data-import-and-checks}).
}
\examples{
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

\dontrun{
# parameters only
readMWRresultsview(respth, columns = 'Characteristic Name')

# parameters and units
readMWRresultsview(respth, columns = c('Characteristic Name', 'Result Unit'))

# all columns
readMWRresultsview(respth)
}
}
