% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamsMWR.R
\docType{data}
\name{streamsMWR}
\alias{streamsMWR}
\title{A simple features object of streams in Massachusetts}
\format{
A simple features object of linestring geometries with one attribute called \code{dLevel} specifying the level of detail.
}
\usage{
streamsMWR
}
\description{
A simple features object of streams in Massachusetts
}
\details{
All geometries are simplified using a tolerance of ten meters.
}
\examples{
\dontrun{
library(sf)
library(dplyr)

streamsMWR <- st_read('~/Desktop/NHD/NHDFlowline_fcode_46006_vis_101k_noattr.shp') \%>\% 
 st_zm() \%>\% 
 st_simplify(dTolerance = 10, preserveTopology = TRUE) \%>\% 
 st_transform(crs = 4326) \%>\% 
 select(dLevel)

save(streamsMWR, file = 'data/streamsMWR.RData', compress = 'xz')
}
}
\keyword{datasets}
