% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudburyMWR.R
\docType{data}
\name{sudburyMWR}
\alias{sudburyMWR}
\title{A simple features object of the Sudbury River watershed}
\format{
A simple features object of linestring geometries
}
\usage{
sudburyMWR
}
\description{
A simple features object of the Sudbury River watershed
}
\examples{
\dontrun{
library(sf)
library(dplyr)

sudburyMWR <- st_read(
  '~/Desktop/Sudbury_watershed_singleboundary/Sudbury_watershed_singleboundary.shp'
 ) \%>\% 
 st_zm()

save(sudburyMWR, file = 'data/subdburyMWR.RData', compress = 'xz')
}
}
\keyword{datasets}
