% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabMWRwqx.R
\name{tabMWRwqx}
\alias{tabMWRwqx}
\title{Create and save tables in a single workbook for WQX upload}
\usage{
tabMWRwqx(
  res = NULL,
  acc = NULL,
  sit = NULL,
  wqx = NULL,
  fset = NULL,
  output_dir = NULL,
  output_file = NULL
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{acc}{character string of path to the data quality objectives file for accuracy or \code{data.frame} returned by \code{\link{readMWRacc}}}

\item{sit}{character string of path to the site metadata file or \code{data.frame} for site metadata returned by \code{\link{readMWRsites}}}

\item{wqx}{character string of path to the wqx metadata file or \code{data.frame} for wqx metadata returned by \code{\link{readMWRwqx}}}

\item{fset}{optional list of inputs with elements named \code{res}, \code{acc}, \code{frecom}, \code{sit}, or \code{wqx} overrides the other arguments}

\item{output_dir}{character string of the output directory for the results, default is the working directory}

\item{output_file}{optional character string for the file name, must include .xlsx suffix}
}
\value{
An Excel workbook is saved to location defined in \code{output_dir} with the name \code{output_file} as default
}
\description{
Create and save tables in a single workbook for WQX upload
}
