% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdMWR.R
\docType{data}
\name{thresholdMWR}
\alias{thresholdMWR}
\title{Master thresholds list for analysis of results data}
\format{
A \code{data.frame} of 39 rows and 6 columns
}
\usage{
thresholdMWR
}
\description{
Master thresholds list for analysis of results data
}
\details{
This information is used to verify the correct format of input data and for minor formatting of output data for upload to WQX.  A column showing the corresponding WQX names is also included.
}
\examples{
\dontrun{
library(readxl)
library(dplyr)

thresholdMWR <- read_excel('inst/extdata/ThresholdMapping.xlsx') \%>\% 
  rename(
  `Simple Parameter` = Simple_Parameter,
  uom = UOM
  )

save(thresholdMWR, file = 'data/threhsoldMWR.RData', compress = 'xz')
}
}
\keyword{datasets}
