% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilMWRdaterange.R
\name{utilMWRdaterange}
\alias{utilMWRdaterange}
\title{Filter results data by date range}
\usage{
utilMWRdaterange(resdat, dtrng = NULL)
}
\arguments{
\item{resdat}{results data as returned by \code{\link{readMWRresults}}}

\item{dtrng}{character string of length two for the date ranges as YYYY-MM-DD}
}
\value{
\code{resdat} filtered by \code{dtrng}, otherwise \code{resdat} unfiltered if \code{dtrng = NULL}
}
\description{
Filter results data by date range
}
\examples{
# results file path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# results data
resdat <- readMWRresults(respth)

# filtered results
utilMWRdaterange(resdat, dtrng = c('2021-06-01', '2021-06-30'))
}
