% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilMWRinput.R
\name{utilMWRinput}
\alias{utilMWRinput}
\title{Utility function to import data as paths or data frames}
\usage{
utilMWRinput(res, frecom, runchk = TRUE)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{frecom}{character string of path to the data quality objectives file for frequency and completeness or \code{data.frame} returned by \code{\link{readMWRfrecom}}}

\item{runchk}{logical to run data checks with \code{\link{checkMWRresults}} and \code{\link{checkMWRfrecom}}, applies only if \code{res} or \code{frecom} are file paths}
}
\value{
A two element list with the imported results and data quality objectives, named \code{"resdat"} and \code{"frecomdat"}, respectively.
}
\description{
Utility function to import data as paths or data frames
}
\details{
The function is used internally by others to import data from paths to the relevant files or as data frames returned by \code{\link{readMWRresults}} and \code{\link{readMWRfrecom}}.  For the former, the full suite of data checks can be evaluated with \code{runkchk = T} (default) or suppressed with \code{runchk = F}.
}
\examples{
##
# using file paths

# results path
respth <- system.file('extdata/ExampleResults_final.xlsx', package = 'MassWateR')

# frequency and completeness path
frecompth <- system.file('extdata/ExampleDQOFrequencyCompleteness_final.xlsx', 
     package = 'MassWateR')

inp <- utilMWRinput(res = respth, frecom = frecompth)
inp$resdat
inp$frecomdat

##
# using data frames

# results data
resdat <- readMWRresults(respth)

# frequency and completeness data
frecomdat <- readMWRfrecom(frecompth)

inp <- utilMWRinput(res = respth, frecom = frecompth)
inp$resdat
inp$frecomdat
}
