% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlzMWRmap.R
\name{anlzMWRmap}
\alias{anlzMWRmap}
\title{Analyze results with maps}
\usage{
anlzMWRmap(
  res,
  param,
  acc,
  sit,
  site = NULL,
  resultatt = NULL,
  locgroup = NULL,
  dtrng = NULL,
  ptsize = 4,
  repel = TRUE,
  labsize = 3,
  palcol = "Greens",
  yscl = c("auto", "log", "linear"),
  crs = 4326,
  zoom = 11,
  addwater = "nhd",
  watercol = "lightblue",
  dLevel = "medium",
  maptype = NULL,
  buffdist = 0.02,
  northloc = "tl",
  scaleloc = "br",
  latlon = TRUE,
  ttlsize = 1.2,
  runchk = TRUE,
  warn = TRUE
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{param}{character string of the parameter to plot, must conform to entries in the \code{"Simple Parameter"} column of \code{\link{paramsMWR}}}

\item{acc}{character string of path to the data quality objectives file for accuracy or \code{data.frame} returned by \code{\link{readMWRacc}}}

\item{sit}{character string of path to the site metadata file or \code{data.frame} of site metadata returned by \code{\link{readMWRsites}}}

\item{site}{character string of sites to include, default all}

\item{resultatt}{character string of result attributes to plot, default all}

\item{locgroup}{character string of location groups to plot from the \code{"Location Group"} column in the site metadata file, default all}

\item{dtrng}{character string of length two for the date ranges as YYYY-MM-DD, default all}

\item{ptsize}{numeric for size of the points, use a negative value to omit the points}

\item{repel}{logical indicating if overlapping site labels are offset}

\item{labsize}{numeric for size of the site labels}

\item{palcol}{character string indicating the color palette to be used from \href{https://r-graph-gallery.com/38-rcolorbrewers-palettes.html}{RColorBrewer}, see details}

\item{yscl}{character indicating one of \code{"auto"} (default), \code{"log"}, or \code{"linear"}, see details}

\item{crs}{numeric as a four-digit EPSG number for the coordinate reference system, see details}

\item{zoom}{numeric indicating resolution of the base map, see details}

\item{addwater}{character string as \code{"nhd"} or \code{"osm"} (or \code{NULL} to suppress) to include water bodies as lines or polygons added to the map, see details}

\item{watercol}{character string of color for water objects if \code{addwater = "nhd"} or \code{addwater = "osm"}}

\item{dLevel}{character string indicating \code{"low"}, \code{"medium"} (default), or \code{"high"} for the level of detail if \code{addwater = "nhd"}, see details}

\item{maptype}{character string for the base map type, see details}

\item{buffdist}{numeric for buffer around the bounding box for the selected sites, see details}

\item{northloc}{character string indicating location of the north arrow, see details}

\item{scaleloc}{character string indicating location of the scale bar, see details}

\item{latlon}{logical to include latitude and longitude labels on the plot, default \code{TRUE}}

\item{ttlsize}{numeric value indicating font size of the title relative to other text in the plot}

\item{runchk}{logical to run data checks with \code{\link{checkMWRresults}}, \code{\link{checkMWRacc}}, or \code{\link{checkMWRsites}}, applies only if \code{res}, \code{acc}, or \code{sit} are file paths}

\item{warn}{logical to return warnings to the console (default)}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified.
}
\description{
Analyze results with maps
}
\details{
This function creates a map of summarized results for a selected parameter at each monitoring site.  By default, all dates for the parameter are averaged. Options to filter by site, date range, and result attribute are provided.  Only sites with spatial information in the site metadata file are plotted and a warning is returned for those that do not have this information. The site labels are also plotted next to each point.  The labels can be suppressed by setting \code{labsize = NULL}.

Any acceptable color palette from \href{https://r-graph-gallery.com/38-rcolorbrewers-palettes.html}{RColorBrewer} can be used for \code{palcol}, which is passed to the \code{palette} argument in \code{\link[ggplot2]{scale_fill_distiller}}. These could include any of the sequential color palettes, e.g., \code{"Greens"}, \code{"Blues"}, etc.  The diverging and qualitative palettes will also work, but may return uninterpretable color scales.

The default value for \code{crs} is EPSG 4326 for the WGS 84 projection in decimal degrees.  The \code{crs} argument is passed to \code{\link[sf]{st_as_sf}} and any acceptable CRS appropriate for the data can be used.

The results shown on the map represent the parameter average for each site within the date range provided by \code{dtrng}.  The average may differ depending on the value provided to the \code{yscl} argument.  Log10-distributed parameters use the geometric mean and normally-distributed parameters use the arithmetic mean.  The distribution is determined from the \code{ycsl} argument. If \code{yscl = "auto"} (default), the distribution is determined automatically from the data quality objective file for accuracy, i.e., parameters with "log" in any of the columns are summarized with the geometric mean, otherwise arithmetic. Setting \code{yscl = "linear"} or \code{yscl = "log"} will the use arithmetic or geometric summaries, respectively, regardless of the information in the data quality objective file for accuracy.

Using \code{addwater = "nhd"} (default) will include lines and polygons of natural water bodies defined using the National Hydrography Dataset (NHD) and included as data files with the package as \code{\link{pondsMWR}}, \code{\link{riversMWR}}, and \code{\link{streamsMWR}}. The level of detail for the water bodies is defined by \code{dLevel} as \code{"low"}, \code{"medium"} (default), or \code{"high"}. The detail level only applies for the NHD data.

Using \code{addwater = "osm"} will include lines and polygons of natural water bodies from the OpenStreetMap (\href{https://www.openstreetmap.org/}{OSM}) project, downloaded using the \href{https://docs.ropensci.org/osmdata}{osmdata} package. Using \code{addwater = "nhd"} may be preferred for more accurate display of water bodies in Massachusetts, whereas \code{addwater = "osm"} will include water bodies outside of the state. Use \code{addwater = NULL} to suppress both.

A base map can be plotted using the \code{maptype} argument and is obtained from the \code{\link[ggmap]{get_stamenmap}} function of ggmap.  The \code{zoom} value specifies the resolution of the map.  Use higher values to download map tiles with greater resolution, although this increases the download time.  The \code{maptype} argument describes the type of base map to download. Acceptable options include \code{"terrain"}, \code{"terrain-background"}, \code{"terrain-labels"}, \code{"terrain-lines"}, \code{"toner"}, \code{"toner-2010"}, \code{"toner-2011"}, \code{"toner-background"}, \code{"toner-hybrid"}, \code{"toner-labels"}, \code{"toner-lines"}, \code{"toner-lite"}, or \code{"watercolor"}. Use \code{maptype = NULL} to suppress the base map.

The area around the summarized points can be increased or decreased using the \code{buffdist} argument.  This creates a buffered area around the bounding box for the points, where the units are degrees.

A north arrow and scale bar are also placed on the map as defined by the \code{northloc} and \code{scaleloc} arguments.  The placement for both can be chosen as \code{"tl"}, \code{"tr"}, \code{"bl"}, or \code{"br"} for top-left, top-right, bottom-left, or bottom-right respectively.  Setting either of the arguments to \code{NULL} will suppress the placement on the map.
}
\examples{
# results data path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# results data
resdat <- readMWRresults(respth)

# accuracy path
accpth <- system.file('extdata/ExampleDQOAccuracy.xlsx', package = 'MassWateR')

# accuracy data
accdat <- readMWRacc(accpth)

# site data path
sitpth <- system.file('extdata/ExampleSites.xlsx', package = 'MassWateR')

# site data
sitdat <- readMWRsites(sitpth)

# map with NHD water bodies
anlzMWRmap(res = resdat, param = 'DO', acc = accdat, sit = sitdat, addwater = "nhd")

# map with OpenStreetMap water bodies
anlzMWRmap(res = resdat, param = 'DO', acc = accdat, sit = sitdat, addwater = "osm")

# map without water bodies, base map included
anlzMWRmap(res = resdat, param = 'DO', acc = accdat, sit = sitdat, maptype = 'terrain', 
  addwater = NULL)

}
