% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlzMWRoutlier.R
\name{anlzMWRoutlier}
\alias{anlzMWRoutlier}
\title{Analyze outliers in results file}
\usage{
anlzMWRoutlier(
  res,
  param,
  type = c("month", "site"),
  jitter = TRUE,
  repel = TRUE,
  outliers = FALSE,
  runchk = TRUE,
  warn = TRUE
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{param}{character string of the parameter to plot, must conform entries in the \code{"Simple Parameter"} of \code{\link{paramsMWR}}}

\item{type}{character indicating whether the summaries are grouped by month (default) or site}

\item{jitter}{logical indicating of non-outlier points are jittered over the boxplots}

\item{repel}{logical indicating if overlapping outlier labels are offset}

\item{outliers}{logical indicating if outliers are returned to the console instead of plotting}

\item{runchk}{logical to run data checks with \code{\link{checkMWRresults}}, \code{\link{checkMWRacc}}, \code{\link{checkMWRfrecom}}, applies only if \code{res}, \code{acc}, or \code{frecom} are file paths}

\item{warn}{logical to return warnings to the console (default)}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further modified if \code{outliers = TRUE}, otherwise a data frame of outliers is returned.
}
\description{
Analyze outliers in results file
}
\details{
Outliers are defined following the standard \code{\link[ggplot2]{ggplot}} definition as 1.5 times the inter-quartile range of each boxplot.  The data frame returned if \code{outliers = TRUE} may vary based on the boxplot groupings defined by \code{type}.
}
\examples{
# results data path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# results data
resdat <- readMWRresults(respth)

# outliers by month
anlzMWRoutlier(res = resdat, param = 'DO', type = 'month')

# outliers by site
anlzMWRoutlier(res = resdat, param = 'DO', type = 'site')

# data frame output
anlzMWRoutlier(res = resdat, param = 'DO', type = 'month', outliers = TRUE)
}
