% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formMWRacc.R
\name{formMWRacc}
\alias{formMWRacc}
\title{Format data quality objective accuracy data}
\usage{
formMWRacc(accdat)
}
\arguments{
\item{accdat}{input data fram}
}
\description{
Format data quality objective accuracy data
}
\details{
This function is used internally within \code{\link{readMWRacc}} to format the input data for downstream analysis.  The formatting includes:

\itemize{
\item{Convert Parameter: }{All parameters are converted to \code{Simple Parameter} in \code{\link{paramsMWR}} as needed}
}
}
\examples{
accpth <- system.file('extdata/ExampleDQOAccuracy_final.xlsx', package = 'MassWateR')

accdat <- readxl::read_excel(accpth, na = c('NA', 'na', '')) 

formMWRacc(accdat)
}
