% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_completeness.R
\name{qc_completeness}
\alias{qc_completeness}
\title{Run quality control completeness checks for water quality monitoring results}
\usage{
qc_completeness(res, frecom, runchk = TRUE, warn = TRUE)
}
\arguments{
\item{res}{character string of path to eresults file or \code{data.frame} for results returned by \code{\link{read_results}}}

\item{frecom}{character string of path to the data quality objectives file for completeness or \code{data.frame} returned by \code{\link{read_frecom}}}

\item{runchk}{logical to run data checks with \code{\link{check_results}} and \code{\link{check_frecom}}, applies only if \code{res} or \code{frecom} are file paths}

\item{warn}{logical to return warnings to the console (default)}
}
\value{
A summarized \code{data.frame} of completeness results for quality control in long format.  Each row applies to a specific completeness check for a parameter. The total observations (\code{obs}) for each parameter are also shown, which is repeated for a parameter across the rows to calculate percentages.  The \code{check} column shows the relevant activity type that is assessed for completeness, the \code{count} columns shows the number of observations that apply to the check, the \code{standard} shows the relevant percentage required for quality control check from the quality control objectives file, the \code{percent} column shows the calculated percent taken from the input data, and the \code{met} column shows if the standard was met by comparing if \code{percent} is greater than or equal to \code{standard}.
}
\description{
Run quality control completeness checks for water quality monitoring results
}
\details{
The function can be used with inputs as paths to the relevant files or as data frames returned by \code{\link{read_results}} and \code{\link{read_frecom}}.  For the former, the full suite of data checks can be evaluated with \code{runkchk = T} (default) or suppressed with \code{runchk = F}.  In the latter case, downstream analyses may not work if data are formatted incorrectly.

Note that completeness is only evaluated on parameters in the \code{Parameter} column in the data quality objectives completeness file.  A warning is returned if there are parameters in that column that are not found in the results file.
}
\examples{
##
# using file paths

# results path
respth <- system.file('extdata/ExampleResults_final.xlsx', package = 'MassWateR')

# completeness path
frecompth <- system.file('extdata/ExampleDQOFrequencyCompleteness_final.xlsx', 
     package = 'MassWateR')

qc_completeness(res = respth, frecom = frecompth)

##
# using data frames

# results data
resdat <- read_results(respth)

# completeness data
frecomdat <- read_frecom(frecompth)

qc_completeness(res = resdat, frecom = frecomdat)

}
