% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dqocompleteness.R
\name{read_dqocompleteness}
\alias{read_dqocompleteness}
\title{Read data quality objectives for completeness from an external file}
\usage{
read_dqocompleteness(pth, runchk = TRUE)
}
\arguments{
\item{pth}{character string of path to the file}

\item{runchk}{logical to run data checks with \code{\link{check_dqocompleteness}}}
}
\value{
A formatted data frame of data quality objectives for completeness that can be used for downstream analysis
}
\description{
Read data quality objectives for completeness from an external file
}
\details{
Data are imported with \code{\link[readxl]{read_excel}} and checked with \code{\link{check_dqocompleteness}}.
}
\examples{
pth <- system.file('extdata/ExampleDQOCompleteness_final.xlsx', package = 'MassWateR')

dqocomplete <- read_dqocompleteness(pth)
head(dqocomplete)
}
