% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sites.R
\name{read_sites}
\alias{read_sites}
\title{Read site metadata from an external file}
\usage{
read_sites(sitpth, runchk = TRUE)
}
\arguments{
\item{sitpth}{character string of path to the site metadata file}

\item{runchk}{logical to run data checks with \code{\link{check_sites}}}
}
\value{
A formatted data frame of site metadata that can be used for downstream analysis
}
\description{
Read site metadata from an external file
}
\details{
Data are imported with \code{\link[readxl]{read_excel}} and checked with \code{\link{check_sites}}.
}
\examples{
sitpth <- system.file('extdata/ExampleSites_final.xlsx', package = 'MassWateR')

sitdat <- read_sites(sitpth)
head(sitdat)
}
