% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riversMWR.R
\docType{data}
\name{riversMWR}
\alias{riversMWR}
\title{A simple features object of rivers in Massachusetts}
\format{
A simple features object of linestring geometries with one attribute called \code{dLevel} specifying the level of detail.
}
\usage{
riversMWR
}
\description{
A simple features object of rivers in Massachusetts
}
\details{
All geometries are simplified using a tolerance of ten meters.
}
\examples{
\dontrun{
library(sf)
library(dplyr)

riversMWR <- st_read('~/Desktop/NHD/NHDArea_noattr.shp') \%>\% 
 st_make_valid() \%>\% 
 st_simplify(dTolerance = 10, preserveTopology = TRUE) \%>\% 
 select(dLevel)

save(riversMWR, file = 'data/riversMWR.RData', compress = 'xz')
}
}
\keyword{datasets}
