% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riversMWR.R
\docType{data}
\name{riversMWR}
\alias{riversMWR}
\title{A geometry set of rivers in Massachusetts}
\format{
A geometry set of linestring geometries
}
\usage{
riversMWR
}
\description{
A geometry set of rivers in Massachusetts
}
\examples{
\dontrun{
library(sf)
library(dplyr)

riversMWR <- st_read('~/Desktop/NHD/NHDArea_noattr.shp') \%>\% 
 st_make_valid() \%>\% 
 st_simplify(dTolerance = 10, preserveTopology = TRUE) \%>\% 
 st_transform(crs = 4326) \%>\% 
 st_geometry()

save(riversMWR, file = 'data/riversMWR.RData', compress = 'xz')
}
}
\keyword{datasets}
