% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_completeness.R
\name{tab_completeness}
\alias{tab_completeness}
\title{Create a formatted table of quality control completeness checks}
\usage{
tab_completeness(
  res,
  dqocom,
  runchk = TRUE,
  warn = TRUE,
  type = c("summary", "percent"),
  pass_col = "green",
  fail_col = "red",
  digits = 0,
  suffix = "\%"
)
}
\arguments{
\item{res}{character string of path to eresults file or \code{data.frame} for results returned by \code{\link{read_results}}}

\item{dqocom}{character string of path to the data quality objectives file for completeness or \code{data.frame} returned by \code{\link{read_dqocompleteness}}}

\item{runchk}{logical to run data checks with \code{\link{check_results}} and \code{\link{check_dqocompleteness}}, applies only if \code{res} or \code{dqocom} are file paths}

\item{warn}{logical to return warnings to the console (default)}

\item{type}{character string indicating if a summary or percentage table is returned}

\item{pass_col}{character string for the cell color of checks that pass, applies only if \code{type = 'percent'}}

\item{fail_col}{character string for the cell color of checks that fail, applies only if \code{type = 'percent'}}

\item{digits}{numeric indicating number of significant digits to report for percentages}

\item{suffix}{character string indicating suffix to append to percentage values}
}
\value{
A \code{\link{flextable}} object with formatted results showing summary counts if \code{type = 'summary'} or percentage observations if \code{type = 'percent'} for all completeness checks for each parameter.
}
\description{
Create a formatted table of quality control completeness checks
}
\details{
Inputs for the results and data quality objectives for completeness are processed internally with \code{\link{qc_completeness}} and the same arguments are accepted for this function, in addition to others listed above.
}
\examples{

# results path
respth <- system.file('extdata/ExampleResults_final.xlsx', package = 'MassWateR')

# completeness path
dqocompth <- system.file('extdata/ExampleDQOCompleteness_final.xlsx', package = 'MassWateR')

## 
# summary table 

tab_completeness(res = respth, dqocom = dqocompth, type = 'summary')

## 
# percent table 

tab_completeness(res = respth, dqocom = dqocompth, type = 'percent')

}
