% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabMWRacc.R
\name{tabMWRacc}
\alias{tabMWRacc}
\title{Create a formatted table of quality control accuracy checks}
\usage{
tabMWRacc(
  res,
  acc,
  runchk = TRUE,
  warn = TRUE,
  accchk = c("Field Blanks", "Lab Blanks", "Field Duplicates", "Lab Duplicates",
    "Lab Spikes", "Instrument Checks (post sampling)"),
  type = c("summary", "percent"),
  pass_col = "green",
  fail_col = "red",
  digits = 0,
  suffix = "\%"
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{acc}{character string of path to the data quality objectives file for accuracy or \code{data.frame} returned by \code{\link{readMWRacc}}}

\item{runchk}{logical to run data checks with \code{\link{checkMWRresults}} and \code{\link{checkMWRacc}}, applies only if \code{res} or \code{acc} are file paths}

\item{warn}{logical to return warnings to the console (default)}

\item{accchk}{character string indicating which accuracy check to return, one to any of \code{"Field Blanks"}, \code{"Lab Blanks"}, \code{"Field Duplicates"}, \code{"Lab Duplicates"}, \code{"Lab Spikes"}, or \code{"Instrument Checks (post sampling)"}}

\item{type}{character string indicating \code{summary} or \code{percent} tabular output, see datails}

\item{pass_col}{character string for the cell color of checks that pass, applies only if \code{type = 'percent'}}

\item{fail_col}{character string for the cell color of checks that fail, applies only if \code{type = 'percent'}}

\item{digits}{numeric indicating number of significant digits to report for percentages}

\item{suffix}{character string indicating suffix to append to percentage values}
}
\value{
A \code{\link{flextable}} object with formatted results.
}
\description{
Create a formatted table of quality control accuracy checks
}
\details{
The function can be used with inputs as paths to the relevant files or as data frames returned by \code{\link{readMWRresults}} and \code{\link{readMWRacc}}.  For the former, the full suite of data checks can be evaluated with \code{runkchk = T} (default) or suppressed with \code{runchk = F}, as explained in the relevant help files.  In the latter case, downstream analyses may not work if data are formatted incorrectly. Also note that accuracy is only evaluated on parameters that are shared between the results file and data quality objectives file for accuracy.  A warning is returned for parameters that do not match between the files. This warning can be suppressed by setting \code{warn = FALSE}.

The quality control tables for accuracy show ........

Inputs for the results and data quality objectives for accuracy are processed internally with \code{\link{qcMWRacc}} and the same arguments are accepted for this function, in addition to others listed above.
}
\examples{

##
# using file paths

# results path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# accuracy path
accpth <- system.file('extdata/ExampleDQOAccuracy.xlsx', 
     package = 'MassWateR')

# table as summary
tabMWRacc(res = respth, acc = accpth, type = 'summary', accchk = 'Field Blanks')

# table as percent
tabMWRacc(res = respth, acc = accpth, type = 'percent')

##
# using data frames

# results data
resdat <- readMWRresults(respth)

# accuracy data
accdat <- readMWRacc(accpth)

# table as summary
tabMWRacc(res = resdat, acc = accdat, type = 'summary', accchk = 'Field Blanks')

# table as percent
tabMWRacc(res = resdat, acc = accdat, type = 'percent')
}
