% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabMWRacc.R
\name{tabMWRacc}
\alias{tabMWRacc}
\title{Create a formatted table of quality control accuracy checks}
\usage{
tabMWRacc(
  res,
  acc,
  runchk = TRUE,
  warn = TRUE,
  accchk = c("Field Blanks", "Lab Blanks", "Field Duplicates", "Lab Duplicates",
    "Lab Spikes", "Instrument Checks"),
  type = c("individual", "summary", "percent"),
  pass_col = "green",
  fail_col = "red",
  frecom = NULL,
  suffix = "\%",
  caption = TRUE
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{acc}{character string of path to the data quality objectives file for accuracy or \code{data.frame} returned by \code{\link{readMWRacc}}}

\item{runchk}{logical to run data checks with \code{\link{checkMWRresults}} and \code{\link{checkMWRacc}}, applies only if \code{res} or \code{acc} are file paths}

\item{warn}{logical to return warnings to the console (default)}

\item{accchk}{character string indicating which accuracy check to return, one to any of \code{"Field Blanks"}, \code{"Lab Blanks"}, \code{"Field Duplicates"}, \code{"Lab Duplicates"}, \code{"Lab Spikes"}, or \code{"Instrument Checks"}}

\item{type}{character string indicating \code{individual}, \code{summary} or \code{percent} tabular output, see details}

\item{pass_col}{character string for the cell color of checks that pass, applies only if \code{type = 'percent'}}

\item{fail_col}{character string for the cell color of checks that fail, applies only if \code{type = 'percent'}}

\item{frecom}{character string of path to the data quality objectives file for frequency and completeness or \code{data.frame} returned by \code{\link{readMWRfrecom}}, applies only if \code{type = "summary"} or \code{type = "percent"}}

\item{suffix}{character string indicating suffix to append to percentage values}

\item{caption}{logical to include a caption from \code{accchk}, only applies if \code{type = "individual"}}
}
\value{
A \code{\link{flextable}} object with formatted results.
}
\description{
Create a formatted table of quality control accuracy checks
}
\details{
The function can be used with inputs as paths to the relevant files or as data frames returned by \code{\link{readMWRresults}} and \code{\link{readMWRacc}}.  For the former, the full suite of data checks can be evaluated with \code{runkchk = T} (default) or suppressed with \code{runchk = F}, as explained in the relevant help files.  In the latter case, downstream analyses may not work if data are formatted incorrectly. Also note that accuracy is only evaluated on parameters that are shared between the results file and data quality objectives file for accuracy.  A warning is returned for parameters that do not match between the files. This warning can be suppressed by setting \code{warn = FALSE}.

The function can return three types of tables as specified with the \code{type} argument: \code{"individual"}, \code{"summary"}, or \code{"percent"}.  The individual tables are specific to each type of accuracy check for each parameter (e.g., field blanks, lab blanks, etc.).  The summary table summarizes all accuracy checks by the number of checks and how many hit/misses are returned for each across all parameters.  The percent table is similar to the summary table, but showing only percentages with appropriate color-coding for hit/misses. The data quality objectives file for frequency and completeness is required if \code{type = "summary"} or \code{type = "percent"}.

For \code{type = "individual"}, the quality control tables for accuracy are retrieved by specifying the check with the \code{accchk} argument.  The \code{accchk} argument can be used to specify one of the following values to retrieve the relevant tables: \code{"Field Blanks"}, \code{"Lab Blanks"}, \code{"Field Duplicates"}, \code{"Lab Duplicates"}, \code{"Lab Spikes"}, or \code{"Instrument Checks"}.

For \code{type = "summary"}, the function summarizes all accuracy checks by counting the number of quality control checks, number of misses, and percent acceptance for each parameter. All accuracy checks are used and the \code{accchk} argument does not apply.

For \code{type = "percent"}, the function returns a similar table as for the summary option, except only the percentage of checks that pass for each parameter are shown in wide format. Cells are color-coded based on the percentage of checks that have passed using the percent thresholds from the \code{\% Completeness} column of the data quality objectives file for frequency and completeness. Parameters without an entry for \code{\% Completeness} are not color-coded and an appropriate warning is returned. All accuracy checks are used and the \code{accchk} argument does not apply.

Inputs for the results and data quality objectives for accuracy are processed internally with \code{\link{qcMWRacc}} and the same arguments are accepted for this function, in addition to others listed above.
}
\examples{

##
# using file paths

# results path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# accuracy path
accpth <- system.file('extdata/ExampleDQOAccuracy.xlsx', 
     package = 'MassWateR')

# frequency and completeness path, only needed if type = "percent"
frecompth <- system.file('extdata/ExampleDQOFrequencyCompleteness.xlsx', 
     package = 'MassWateR')

# table as individual
tabMWRacc(res = respth, acc = accpth, type = 'individual', accchk = 'Field Blanks')

# table as summary
tabMWRacc(res = respth, acc = accpth, type = 'summary', frecom = frecompth)

# table as percent
tabMWRacc(res = respth, acc = accpth, type = 'percent', frecom = frecompth)

##
# using data frames

# results data
resdat <- readMWRresults(respth)

# accuracy data
accdat <- readMWRacc(accpth)

# frequency and completeness data, only needed if type = "percent"
frecomdat <- readMWRfrecom(frecompth)

# table as individual
tabMWRacc(res = resdat, acc = accdat, type = 'individual', accchk = 'Field Blanks')

# table as summary
tabMWRacc(res = resdat, acc = accdat, type = 'summary', frecom = frecomdat)

# table as percent
tabMWRacc(res = resdat, acc = accdat, type = 'percent', frecom = frecomdat)
}
