% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabMWRfre.R
\name{tabMWRfre}
\alias{tabMWRfre}
\title{Create a formatted table of quality control frequency checks}
\usage{
tabMWRfre(
  res,
  frecom,
  runchk = TRUE,
  warn = TRUE,
  type = c("summary", "percent"),
  pass_col = "green",
  fail_col = "red",
  digits = 0,
  suffix = "\%"
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{frecom}{character string of path to the data quality objectives file for frequency and completeness or \code{data.frame} returned by \code{\link{readMWRfrecom}}}

\item{warn}{logical to return warnings to the console (default)}

\item{type}{character string indicating \code{summary} or \code{percent} tabular output, see datails}

\item{pass_col}{character string for the cell color of checks that pass, applies only if \code{type = 'percent'}}

\item{fail_col}{character string for the cell color of checks that fail, applies only if \code{type = 'percent'}}

\item{digits}{numeric indicating number of significant digits to report for percentages}

\item{suffix}{character string indicating suffix to append to percentage values}
}
\value{
A \code{\link{flextable}} object with formatted results.
}
\description{
Create a formatted table of quality control frequency checks
}
\details{
The function can be used with inputs as paths to the relevant files or as data frames returned by \code{\link{readMWRresults}} and \code{\link{readMWRfrecom}}.  For the former, the full suite of data checks can be evaluated with \code{runkchk = T} (default) or suppressed with \code{runchk = F}, as explained in the relevant help files.  In the latter case, downstream analyses may not work if data are formatted incorrectly.  Also note that frequency is only evaluated on parameters in the \code{Parameter} column in the data quality objectives frequency and completeness file.  A warning is returned if there are parameters in that column that are not found in the results file.  This warning can be suppressed by setting \code{warn = FALSE}.

The quality control tables for frequency show the number of records that apply to a given check (e.g., Lab Blank, Field Blank, etc.) relative to the number of "regular" data records (e.g., field samples or measures) for each parameter.  A summary of all frequency checks for each parameter is provided if \code{type = "summary"} or a color-coded table showing similar information as percentages for each parameter is provided if \code{type = "percent"}.

Inputs for the results and data quality objectives for frequency and completeness are processed internally with \code{\link{qcMWRcom}} and the same arguments are accepted for this function, in addition to others listed above.
}
\examples{

##
# using file paths

# results path
respth <- system.file('extdata/ExampleResults_final.xlsx', package = 'MassWateR')

# frequency and completeness path
frecompth <- system.file('extdata/ExampleDQOFrequencyCompleteness_final.xlsx', 
     package = 'MassWateR')

# table as summary
tabMWRfre(res = respth, frecom = frecompth, type = 'summary')

# table as percent
tabMWRfre(res = respth, frecom = frecompth, type = 'percent')

##
# using data frames

# results data
resdat <- readMWRresults(respth)

# frequency and completeness data
frecomdat <- readMWRfrecom(frecompth)

# table as summary
tabMWRfre(res = resdat, frecom = frecomdat, type = 'summary')

# table as percent
tabMWRfre(res = resdat, frecom = frecomdat, type = 'percent')
}
