% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabMWRwqx.R
\name{tabMWRwqx}
\alias{tabMWRwqx}
\title{Create and save tables in a single workbook for WQX upload}
\usage{
tabMWRwqx(
  res = NULL,
  acc = NULL,
  sit = NULL,
  wqx = NULL,
  fset = NULL,
  output_dir = NULL,
  output_file = NULL,
  warn = TRUE,
  runchk = TRUE
)
}
\arguments{
\item{res}{character string of path to the results file or \code{data.frame} for results returned by \code{\link{readMWRresults}}}

\item{acc}{character string of path to the data quality objectives file for accuracy or \code{data.frame} returned by \code{\link{readMWRacc}}}

\item{sit}{character string of path to the site metadata file or \code{data.frame} for site metadata returned by \code{\link{readMWRsites}}}

\item{wqx}{character string of path to the wqx metadata file or \code{data.frame} for wqx metadata returned by \code{\link{readMWRwqx}}}

\item{fset}{optional list of inputs with elements named \code{res}, \code{acc}, \code{frecom}, \code{sit}, or \code{wqx} overrides the other arguments}

\item{output_dir}{character string of the output directory for the results, default is the working directory}

\item{output_file}{optional character string for the file name, must include .xlsx suffix}

\item{warn}{logical to return warnings to the console (default)}

\item{runchk}{logical to run data checks with \code{\link{checkMWRresults}}, \code{\link{checkMWRacc}}, \code{\link{checkMWRsites}}, \code{\link{checkMWRwqx}}, applies only if \code{res}, \code{acc}, \code{sit}, or \code{wqx} are file paths}
}
\value{
An Excel workbook named \code{wqxtab.xlsx} (or name passed to \code{output_file}) will be saved in the directory specified by \code{output_dir} (default is the working directory)
}
\description{
Create and save tables in a single workbook for WQX upload
}
\examples{
# results data path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# dqo accuracy data path
accpth <- system.file('extdata/ExampleDQOAccuracy.xlsx', package = 'MassWateR')

# site data path
sitpth <- system.file('extdata/ExampleSites.xlsx', package = 'MassWateR')

# wqx data path
wqxpth <- system.file('extdata/ExampleWQX.xlsx', package = 'MassWateR')
# results data
resdat <- readMWRresults(respth)

# accuracy data
accdat <- readMWRacc(accpth)

# site data
sitdat <- readMWRsites(sitpth)

# wqx data
wqxdat <- readMWRwqx(wqxpth)

\dontrun{
# create workbook working directory
tabMWRwqx(res = resdat, acc = accdat, sit = sitdat, wqx = wqxdat)
}
}
