% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdMWR.R
\docType{data}
\name{thresholdMWR}
\alias{thresholdMWR}
\title{Master thresholds list for analysis of results data}
\format{
A \code{data.frame} of 28 rows and 10 columns
}
\usage{
thresholdMWR
}
\description{
Master thresholds list for analysis of results data
}
\details{
This file includes appropriate threshold values of water quality parameters for marine and freshwater environments based on state standards or typical ranges in Massachusetts.
}
\examples{
\dontrun{
library(readxl)
library(dplyr)

thresholdMWR <- read_excel('inst/extdata/ThresholdMapping.xlsx', na = 'NA') \%>\% 
  rename(
    `Simple Parameter` = Simple_Parameter,
    uom = UOM
  ) \%>\% 
  mutate_if(is.numeric, round, 2)

save(thresholdMWR, file = 'data/thresholdMWR.RData', compress = 'xz')
}
}
\keyword{datasets}
