% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilMWRhttpgrace.R
\name{utilMWRhttpgrace}
\alias{utilMWRhttpgrace}
\title{Load external file from remote source, fail gracefully}
\usage{
utilMWRhttpgrace(remote_file)
}
\arguments{
\item{remote_file}{URL of the external file}
}
\value{
The external file as an RData object
}
\description{
Load external file from remote source, fail gracefully
}
\examples{
fl <- 'https://github.com/massbays-tech/MassWateRdata/raw/main/data/streamsMWR.RData'
utilMWRhttpgrace(fl)
utilMWRhttpgrace('http://httpbin.org/status/404')
}
