% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilMWRinputcheck.R
\name{utilMWRinputcheck}
\alias{utilMWRinputcheck}
\title{Check if required inputs are present for a function}
\usage{
utilMWRinputcheck(inputs)
}
\arguments{
\item{inputs}{list of arguments passed from the parent function}
}
\value{
NULL if all inputs are present, otherwise an error message indicating which inputs are missing
}
\description{
Check if required inputs are present for a function
}
\examples{
inputchk <- formals(tabMWRcom)
\dontrun{
utilMWRinputcheck(inputchk)
}
}
