% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilMWRthresh.R
\name{utilMWRthresh}
\alias{utilMWRthresh}
\title{Get threshold lines from thresholdMWR}
\usage{
utilMWRthresh(
  resdat,
  param,
  thresh = c("fresh", "marine", "none"),
  warn = TRUE
)
}
\arguments{
\item{resdat}{results data as returned by \code{\link{readMWRresults}}}

\item{param}{character string to first filter results by a parameter in \code{"Characteristic Name"}}

\item{thresh}{character indicating if relevant freshwater or marine threshold lines are included, one of \code{"fresh"}, \code{"marine"}, or \code{"none"}}

\item{warn}{logical to return warnings to the console (default)}
}
\value{

}
\description{
Get threshold lines from thresholdMWR
}
\examples{
# results file path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# results data
resdat <- readMWRresults(respth)

# get threshold lines
utilMWRthresh(resdat = resdat, param = 'E.coli', thresh = 'fresh')
}
