% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilMWRthresh.R
\name{utilMWRthresh}
\alias{utilMWRthresh}
\title{Get threshold lines from thresholdMWR}
\usage{
utilMWRthresh(resdat, param, thresh = c("fresh", "marine", "none"))
}
\arguments{
\item{resdat}{results data as returned by \code{\link{readMWRresults}}}

\item{param}{character string to first filter results by a parameter in \code{"Characteristic Name"}}

\item{thresh}{character indicating if relevant freshwater or marine threshold lines are included, one of \code{"fresh"}, \code{"marine"}, or \code{"none"}}
}
\value{
A logical value indicating \code{TRUE} if a log10-scale should be used, \code{FALSE} for arithmetic (linear)
}
\description{
Get threshold lines from thresholdMWR
}
\examples{
# results file path
respth <- system.file('extdata/ExampleResults.xlsx', package = 'MassWateR')

# results data
resdat <- readMWRresults(respth)

# get threshold lines
utilMWRthresh(resdat = resdat, param = 'E.coli', thresh = 'fresh')
}
